+++ CHANGELOG.md
@@ -1,5 +1,50 @@
 # Changelog
 
+## v2.20.0 (2020-12-16)
+
+### Features
+
+ * add dataset definition support for processing jobs
+
+### Bug Fixes and Other Changes
+
+ * include workflow integ tests with clarify and debugger enabled
+ * only run DataParallel and EdgePackaging tests in supported regions
+
+### Documentation Changes
+
+ * fix smp code example, add note for CUDA 11 to sdp
+ * adding note about CUDA 11 to SMP. Small title update PyTorch
+
+## v2.19.0 (2020-12-08)
+
+### Features
+
+ * add tensorflow 1.15.4 and 2.3.1 as valid versions
+ * add py36 as valid python version for pytorch 1.6.0
+ * auto-select container version for p4d and smdistributed
+ * add edge packaging job support
+ * Add Clarify Processor, Model Bias, Explainability, and Quality Monitors support. (#494)
+ * add model parallelism support
+ * add data parallelism support (#454) (#511)
+ * support creating and updating profiler in training job (#444) (#526)
+
+### Bug Fixes and Other Changes
+
+ * bump boto3 and smdebug_rulesconfig versions for reinvent and enable data parallel integ tests
+ * run UpdateTrainingJob tests only during allowed secondary status
+ * Remove workarounds and apply fixes to Clarify and MM integ tests
+ * add p4d to smdataparallel supported instances
+ * Mount metadata directory when starting local mode docker container
+ * add integ test for profiler
+ * Re-enable model monitor integration tests.
+
+### Documentation Changes
+
+ * add SageMaker distributed libraries documentation
+ * update documentation for the new SageMaker Debugger APIs
+ * minor updates to doc strings
+
 ## v2.18.0 (2020-12-03)
 
 ### Features